refactor(session): collapse extension host to two modules
extensions/ dissolves into extension-loader.ts (jiti loading, tool conversion,
errors) and extension-source.ts (work source, hooks, key builders). The schema-
validate-then-copy layer dies: @plot/sdk owns the shared shapes, so discover
results pass through by reference with only identity and workspace-absolute
checks (sources are trusted TypeScript per the architecture rules).
