--- packages/session/src/extensions/errors.ts
@@ -1,47 +1,0 @@
-import { logWideEvent } from "@plot/common/observability";
-import { errorMessage } from "@plot/common/primitives";
-import type { MaybePromise } from "@plot/sdk";
-
-export class PlotExtensionSourceError extends Error {
-	override readonly name = "PlotExtensionSourceError";
-	readonly phase: "load" | "config" | "create" | "discover" | "hook";
-	readonly source?: string;
-
-	constructor(input: {
-		readonly phase: PlotExtensionSourceError["phase"];
-		readonly message: string;
-		readonly source?: string;
-	}) {
-		super(input.message);
-		this.phase = input.phase;
-		if (input.source !== undefined) this.source = input.source;
-	}
-}
-
-export const runMaybePromise = async <A>(
-	phase: PlotExtensionSourceError["phase"],
-	source: string | undefined,
-	thunk: () => MaybePromise<A>,
-): Promise<A> => {
-	try {
-		return await Promise.resolve(thunk());
-	} catch (error) {
-		throw new PlotExtensionSourceError({
-			phase,
-			message: errorMessage(error),
-			...(source === undefined ? {} : { source }),
-		});
-	}
-};
-
-export const logHookError = (error: unknown, hook: string, source: string) =>
-	logWideEvent(
-		{
-			operation: "plot_extension.hook",
-			outcome: "error",
-			hook,
-			source_id: source,
-			error: errorMessage(error),
-		},
-		"error",
-	);
